+++ extensions/nunl/stub/src/main/java/nl/nu/performance/api/client/objects/HeaderBlock.java
@@ -1,0 +1,10 @@
+package nl.nu.performance.api.client.objects;
+
+import nl.nu.performance.api.client.interfaces.Block;
+
+public class HeaderBlock extends Block {
+    // returns title
+    public final StyledText component20() {
+        throw new UnsupportedOperationException("Stub");
+    }
+}