+++ extensions/nunl/stub/src/main/java/nl/nu/performance/api/client/objects/StyledText.java
@@ -1,0 +1,7 @@
+package nl.nu.performance.api.client.objects;
+
+public class StyledText  {
+    public final String getText() {
+        throw new UnsupportedOperationException("Stub");
+    }
+}