+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/HideEndScreenSuggestedVideoPatch.java
@@ -1,0 +1,13 @@
+package app.revanced.extension.youtube.patches;
+
+import app.revanced.extension.youtube.settings.Settings;
+
+@SuppressWarnings("unused")
+public final class HideEndScreenSuggestedVideoPatch {
+    /**
+     * Injection point.
+     */
+    public static boolean hideEndScreenSuggestedVideo() {
+        return Settings.HIDE_END_SCREEN_SUGGESTED_VIDEO.get();
+    }
+}